+++ src/main/java/dndbuilder/view/main_gui/CharacterInfoPanel.java
@@ -1,0 +1,10 @@
+package dndbuilder.view.main_gui;
+
+import javax.swing.JPanel;
+import java.awt.BorderLayout;
+
+public class CharacterInfoPanel extends JPanel {
+    public CharacterInfoPanel(String[] characterShortInfo) {
+        super(new BorderLayout());
+    }
+}